refactor(i18n-tools): separate catalog adapters
